Better syntax

+++ ISO_maker.sh
@@ -18,80 +18,50 @@
 
 # Updating pacman-config + backup of existing
 
-  check_sudo="$(pacman -Qs --color always "sudo" | grep "local" | grep "sudo ")"
-  if [[ "$(pacman -Qs opendoas)" ]] && [[ -z "${check_sudo}" ]]; then
-    su_command="doas"
-  else
-    su_command="sudo"
-  fi
-  if [[ -z "$(pacman -Qs artix-archlinux-support)" ]]; then
-    "$su_command" pacman -Syy --noconfirm artix-archlinux-support
-    "$su_command" pacman-key --populate archlinux
-  fi
-
+  if [[ "$(pacman -Qs opendoas)" ]] && ! [[ "$(pacman -Qs --color always "sudo" | grep "local" | grep "sudo ")" == "" ]]; then COMMAND="doas"; 
+  else SU_COMMAND="sudo"; fi
+  if [[ -z "$(pacman -Qs artix-archlinux-support)" ]]; then "$SU_COMMAND" pacman -Syy --noconfirm artix-archlinux-support; "$SU_COMMAND" pacman-key --populate archlinux; fi
 
 #----------------------------------------------------------------------------------------------------------------------------------
 
 # Assigning parameters
 
-  if [[ "/etc/vconsole.conf" ]] && [[ -z "$KEYMAP" ]]; then
-    KEYMAP="$(</etc/vconsole.conf)" # Defaults to local keymap
-    KEYMAP_sorted=${KEYMAP#*=}
-  fi
+  if [[ "/etc/vconsole.conf" ]] && [[ -z "$KEYMAP" ]]; then KEYMAP="$(</etc/vconsole.conf)"; KEYMAP_sorted=${KEYMAP#*=}; fi
 
 #----------------------------------------------------------------------------------------------------------------------------------
 
 # Checking and installing any missing dependencies
 
   if [[ "$(pacman -Qs opendoas)" ]] && [[ -z "${check_sudo}" ]]; then
-    if [[ -f "/usr/bin/sudo" ]]; then
-      doas rm -rf /usr/bin/sudo
-      RESTORE_sudo="true"
-    fi
-    doas pacman --noconfirm -S sudo
-    echo ""$USER" ALL=(ALL:ALL) NOPASSWD: ALL" | doas tee -a /etc/sudoers > /dev/null
-    DELETE_sudo="true"
+    if [[ -f "/usr/bin/sudo" ]]; then doas rm -rf /usr/bin/sudo; RESTORE_sudo="true"; fi
+    doas pacman --noconfirm -S sudo && echo ""$USER" ALL=(ALL:ALL) NOPASSWD: ALL" | doas tee -a /etc/sudoers > /dev/null && DELETE_sudo="true"
   fi
-  if [[ -z "$(pacman -Qs openssl)" ]] && [[ "$ANSWERFILE_path_minimal" || "$ANSWERFILE_path_full" ]]; then
-    sudo pacman --noconfirm -S openssl
-    DELETE_openssl="true"
-  fi
-  if [[ -z "$(pacman -Qs artools)" ]]; then
-    sudo pacman --noconfirm -S artools iso-profiles
-  fi
+  if [[ -z "$(pacman -Qs openssl)" ]] && [[ "$ANSWERFILE_path_minimal" || "$ANSWERFILE_path_full" ]]; then sudo pacman --noconfirm -S openssl; DELETE_openssl="true"; fi
+  if [[ -z "$(pacman -Qs artools)" ]]; then sudo pacman --noconfirm -S artools iso-profiles; fi
 
 #----------------------------------------------------------------------------------------------------------------------------------
 
 # Cleaning conflicting folders
 
-  if [[ -d "/home/$(whoami)/ISO" ]]; then
-    rm -rf /home/$(whoami)/ISO
-  fi
+  if [[ -d "/home/$(whoami)/ISO" ]]; then rm -rf /${home/$(whoami)/ISO:?}; fi
   if [[ -d "/home/$(whoami)/BUILDISO" ]]; then
-    if [[ -d "/home/$(whoami)/BUILDISO/buildiso/base/artix/bootfs" ]]; then
-      sudo umount -l /home/$(whoami)/BUILDISO/buildiso/base/artix/bootfs
-    fi
-    if [[ -d "/home/$(whoami)/BUILDISO/buildiso/base/artix/rootfs" ]]; then
-      sudo umount -l /home/$(whoami)/BUILDISO/buildiso/base/artix/rootfs
-    fi
-    sudo rm -rf /home/$(whoami)/BUILDISO
+    if [[ -d "/home/$(whoami)/BUILDISO/buildiso/base/artix/bootfs" ]]; then sudo umount -l /home/$(whoami)/BUILDISO/buildiso/base/artix/bootfs; fi
+    if [[ -d "/home/$(whoami)/BUILDISO/buildiso/base/artix/rootfs" ]]; then sudo umount -l /home/$(whoami)/BUILDISO/buildiso/base/artix/rootfs; fi
+    sudo rm -rf /${home/$(whoami)/BUILDISO:?}
   fi
 
 #----------------------------------------------------------------------------------------------------------------------------------
 
 # Copies configs and creates folders
 
-  cp -rf artools /home/$(whoami)/.config
-  cp -rf artools-workspace /home/$(whoami)
-  mkdir /home/$(whoami)/{BUILDISO,ISO}
+  cp -rf artools /home/$(whoami)/.config && cp -rf artools-workspace /home/$(whoami) && mkdir /home/$(whoami)/{BUILDISO,ISO}
   sudo sed -i 's/\/usr\/src\/linux\/version/\/usr\/src\/linux-zen\/version/' /usr/bin/buildiso
 
 #----------------------------------------------------------------------------------------------------------------------------------
 
 # Builds the filesystem and applies modifications
 
-  sudo modprobe loop
-  buildiso -p base -x
+  sudo modprobe loop && buildiso -p base -x
   sudo sed -i 's/--noclear/--autologin root --noclear/' /home/$(whoami)/BUILDISO/buildiso/base/artix/rootfs/etc/dinit.d/tty1
   sudo cp scripts/startup_choice.sh /home/$(whoami)/BUILDISO/buildiso/base/artix/rootfs/etc/profile.d/startup_choice.sh
   sudo mkdir /home/$(whoami)/BUILDISO/buildiso/base/artix/rootfs/{scripts,.nothing,.encrypt,.decrypt}
@@ -103,11 +73,9 @@
       date | sha512sum > /home/$(whoami)/.nothing$index/nothing$index.txt
       openssl enc -aes-256-cbc -md sha512 -a -pbkdf2 -iter 100000 -salt -in "$ANSWERFILE_path_base" -out /home/$(whoami)/.nothing$index/encrypt$index.txt -pass file:/home/$(whoami)/.nothing$index/nothing$index.txt
       sudo cp /home/$(whoami)/.nothing$index/{nothing$index.txt,encrypt$index.txt} /home/$(whoami)/BUILDISO/buildiso/base/artix/rootfs/scripts
-      rm -rf /home/$(whoami)/.nothing$index
+      rm -rf /${home/$(whoami)/.nothing$index:?}
       (( index++ )) || true
-    else
-      (( index++ )) || true
-    fi
+    else (( index++ )) || true; fi
   done
   sudo cp configs/wifi_backend.conf /home/$(whoami)/BUILDISO/buildiso/base/artix/rootfs/etc/NetworkManager/conf.d/wifi_backend.conf
   sudo cp configs/{pacman_with_arch.conf,pacman_without_arch.conf} /home/$(whoami)/BUILDISO/buildiso/base/artix/rootfs/
@@ -124,19 +92,11 @@
 
 # Continues building the ISO with auto-cleanup
 
-  buildiso -p base -sc
-  buildiso -p base -bc
-  buildiso -p base -zc
+  buildiso -p base -sc && buildiso -p base -bc && buildiso -p base -zc
   sudo rm -rf /home/$(whoami)/BUILDISO
-  if [[ "$DELETE_openssl" == "true" ]]; then
-    sudo pacman --noconfirm -Rns openssl
-  fi
-  if [[ "$DELETE_sudo" == "true" ]]; then
-    doas pacman --noconfirm -Rns sudo
-  fi
-  if [[ "$RESTORE_sudo" == "true" ]]; then
-    doas ln -s $(which doas) /usr/bin/sudo
-  fi
+  if [[ "$DELETE_openssl" == "true" ]]; then sudo pacman --noconfirm -Rns openssl; fi
+  if [[ "$DELETE_sudo" == "true" ]]; then doas pacman --noconfirm -Rns sudo; fi
+  if [[ "$RESTORE_sudo" == "true" ]]; then doas ln -s $(which doas) /usr/bin/sudo; fi
   echo
   echo "----------------------------------------------------------------"
   echo "------YOUR CUSTOM ISO CAN BE FOUND AT /home/$(whoami)/ISO/base------"
